backend/questions/requirements.txt
python-dotenv
openai
tqdm
supabase
argparse